Move param-parsing code into shared library
