+++ workflow/v1_2/workflow.go
@@ -9,6 +9,7 @@ import (
 	"github.com/antha-lang/antha/laboratory/effects"
 	"github.com/antha-lang/antha/microArch/driver/liquidhandling"
 	"github.com/antha-lang/antha/workflow"
+	"github.com/antha-lang/antha/workflow/migrate"
 )
 
 type workflowv1_2 struct {
@@ -71,75 +72,6 @@ type mixTaskResult struct {
 	TimeEstimate time.Duration
 }
 
-type bakedInFile struct {
-	Name  string `json:"name"`
-	Bytes struct {
-		Bytes []byte `json:"bytes"`
-	} `json:"bytes"`
-}
-
-// maybeMigrateFileParam returns either the original json representations, or a json representation as serialized files, if possible.
-func maybeMigrateFileParam(fm *effects.FileManager, param json.RawMessage) (json.RawMessage, error) {
-
-	if js, err := maybeMigrateFileFlat(fm, param); js != nil {
-		return js, nil
-	} else if err != nil {
-		return nil, err
-	} else if js, err := maybeMigrateFileArray(fm, param); js != nil {
-		return js, nil
-	} else if err != nil {
-		return nil, err
-	} else {
-		return param, nil
-	}
-}
-
-// maybeMigrateFileArray returns a nil json.RawMessage unless: no error was encountered during unmarshalling and every value within the slice unmarshals as a valid wtype.File. returns a non-nil error only if an error was encountered when writing a file value to disk.
-func maybeMigrateFileArray(fm *effects.FileManager, param json.RawMessage) (json.RawMessage, error) {
-	var bifArr []*bakedInFile
-	if err := json.Unmarshal([]byte(param), &bifArr); err != nil || bifArr == nil {
-		return nil, nil
-	}
-
-	js := make([]*wtype.File, len(bifArr))
-	for i, bif := range bifArr {
-		if !bif.hasData() {
-			return nil, nil
-		} else if f, err := bif.moveDataToFile(fm); err != nil {
-			return nil, err
-		} else {
-			js[i] = f
-		}
-	}
-
-	return json.Marshal(js)
-}
-
-// maybeMigrateFileFlat returns a nil json.RawMessage unless: no error was encountered during unmarshalling as a valid wtype.File. returns a non-nil error only if an error was encountered when writing a file value to disk.
-func maybeMigrateFileFlat(fm *effects.FileManager, param json.RawMessage) (json.RawMessage, error) {
-	var bif *bakedInFile
-	if err := json.Unmarshal([]byte(param), &bif); err != nil || !bif.hasData() {
-		return nil, nil
-	} else if f, err := bif.moveDataToFile(fm); err != nil {
-		return nil, err
-	} else {
-		return json.Marshal(f)
-	}
-}
-
-// hasData checks that structure is a valid bakedInFile (rather than an empty struct)
-func (bif *bakedInFile) hasData() bool {
-	return bif != nil && bif.Name != "" && len(bif.Bytes.Bytes) > 0
-}
-
-func (bif *bakedInFile) moveDataToFile(fm *effects.FileManager) (*wtype.File, error) {
-	if f, err := fm.WriteAll(bif.Bytes.Bytes, bif.Name); err != nil {
-		return nil, err
-	} else {
-		return f.AsInput(), nil
-	}
-}
-
 func (wf *workflowv1_2) MigrateElementParameters(fm *effects.FileManager, name string) (workflow.ElementParameterSet, error) {
 	v, pr := wf.Parameters[name]
 	if !pr {
@@ -148,7 +80,7 @@ func (wf *workflowv1_2) MigrateElementParameters(fm *effects.FileManager, name s
 	pset := make(workflow.ElementParameterSet)
 
 	for pname, pval := range v {
-		if param, err := maybeMigrateFileParam(fm, pval); err != nil {
+		if param, err := migrate.MaybeMigrateFileParam(fm, pval); err != nil {
 			return nil, err
 		} else {
 			pset[workflow.ElementParameterName(pname)] = param
